更新鉴权逻辑

+++ api/handler/post.go
@@ -12,6 +12,11 @@ import (
 )
 
 func AddPost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+	role := RightAuth(w, r, p)
+	if role == "user" {
+		sendErrorResponse(w, def.ErrorRequestBodyParseFailed)
+		return
+	}
 	req, _ := ioutil.ReadAll(r.Body)
 	pbody := &def.Post{}
 
@@ -32,6 +37,11 @@ func AddPost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 }
 
 func UpdatePost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+	role := RightAuth(w, r, p)
+	if role != "admin" || role != "editor" {
+		sendErrorResponse(w, def.ErrorRequestBodyParseFailed)
+		return
+	}
 	pid := p.ByName("id")
 	pint, _ := strconv.Atoi(pid)
 	req, _ := ioutil.ReadAll(r.Body)
@@ -52,6 +62,11 @@ func UpdatePost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 }
 
 func DeletePost(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+	role := RightAuth(w, r, p)
+	if role != "admin" || role != "editor" {
+		sendErrorResponse(w, def.ErrorRequestBodyParseFailed)
+		return
+	}
 	pid, _ := strconv.Atoi(p.ByName("id"))
 	err := db.DeletePost(pid)
 	if err != nil {
